+++ src/apps/tdse/tdse.cc
@@ -688,7 +688,7 @@ void propagate(World& world, int step0) {
 
     zero_field_time = 10.0*time_step;
 
-    int nstep = (param.target_time + zero_field_time)/time_step + 1;
+    int nstep = int((param.target_time + zero_field_time)/time_step + 1);
 
     // Ensure everyone has the same data
     world.gop.broadcast(c);
